% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{voice_gravy}
\alias{voice_gravy}
\title{Main function of the voice gravy package}
\usage{
voice_gravy(wav_path, result_type = "tibble")
}
\arguments{
\item{wav_path}{path to wav file that is to be processed. Currently
only uncompressed mono .wav files are supported.}

\item{result_type}{specify type of returned object.
Current values are:
\itemize{
\item tibble (== default)
\item AsspDataObj (the main in-memory object of the wrassp package
which can be saved to disk using the \code{wrassp::write.AsspDataObj())}
function.
}}
}
\value{
object of type specified with \code{result_type}
}
\description{
Bit longer info
}
\details{
the measures currently computed are:
\itemize{
\item F0
\item Formants F1-F4
\item H1(*)
\item H2(*)
\item H4(*)
\item A1(*)
\item A2(*)
\item A3(*)
\item 2K(*)
\item 5K
\item H1(*)-H2(*)
\item H2(*)-H4(*)
\item H1(*)-A1(*)
\item H1(*)-A2(*)
\item H1(*)-A3(*)
\item H4(*)-2K(*)
\item 2K(*)-5K
\item Energy
\item Cepstral Peak Prominence
\item Harmonic to Noise Ratios
\item Subharmonic to Harmonic Ratio
\item Strength of Excitation
}
where (*) indicates that the harmonic/spectral amplitudes
are reported with and without corrects for formant frequencies and bandwidths.
}
